Moved cid-impl to /src/service, updated amp-user-notification.md
