+++ src/services.js
@@ -89,10 +89,10 @@ export function bindForDoc(nodeOrDoc) {
 
 /**
  * @param {!Node|!./service/ampdoc-impl.AmpDoc} nodeOrDoc
- * @return {!Promise<!./cid-impl.Cid>}
+ * @return {!Promise<!./service/cid-impl.Cid>}
  */
 export function cidForDoc(nodeOrDoc) {
-  return /** @type {!Promise<!./cid-impl.Cid>} */ ( // eslint-disable-line max-len
+  return /** @type {!Promise<!./service/cid-impl.Cid>} */ ( // eslint-disable-line max-len
       getElementServiceForDoc(nodeOrDoc, 'cid', 'amp-analytics'));
 }
 
@@ -100,10 +100,10 @@ export function cidForDoc(nodeOrDoc) {
  * Returns a promise for the CID service or a promise for null if the service
  * is not available on the current page.
  * @param {!Node|!./service/ampdoc-impl.AmpDoc} nodeOrDoc
- * @return {!Promise<?./cid-impl.Cid>}
+ * @return {!Promise<?./service/cid-impl.Cid>}
  */
 export function cidForDocOrNull(nodeOrDoc) {
-  return /** @type {!Promise<?./cid-impl.Cid>} */ ( // eslint-disable-line max-len
+  return /** @type {!Promise<?./service/cid-impl.Cid>} */ ( // eslint-disable-line max-len
       getElementServiceIfAvailableForDoc(nodeOrDoc, 'cid', 'amp-analytics'));
 }
